Continue decompiling field_specials

+++ include/field_specials.h
@@ -3,7 +3,7 @@
 
 u8 GetLeadMonIndex(void);
 u8 sub_813B260(void);
-u16 get_unknown_box_id(void);
+u8 get_unknown_box_id(void);
 bool8 InMultiBattleRoom(void);
 
 #endif // GUARD_FIELD_SPECIALS_H
